package rename to openapi_terraform_docs_generator
