+++ utils/terraform_docs_generator/main.go
@@ -1,13 +1,13 @@
 package main
 
 import (
-	"github.com/dikhan/terraform-provider-openapi/utils/terraform_docs_generator/openapi"
+	"github.com/dikhan/terraform-provider-openapi/utils/terraform_docs_generator/openapi_terraform_docs_generator"
 	"log"
 	"os"
 )
 
 func main() {
-	terraformProviderDocGenerator, err := openapi.NewTerraformProviderDocGenerator("openapi", "")
+	terraformProviderDocGenerator, err := openapi_terraform_docs_generator.NewTerraformProviderDocGenerator("openapi", "")
 	if err != nil {
 		log.Fatal(err)
 	}
