+++ utils/terraform_docs_generator/openapi_terraform_docs_generator/openapi_terraform_provider_documentation.go
@@ -1,7 +1,7 @@
-package openapi
+package openapi_terraform_docs_generator
 
 import (
-	"github.com/dikhan/terraform-provider-openapi/utils/terraform_docs_generator/openapi/templates/zendesk"
+	"github.com/dikhan/terraform-provider-openapi/utils/terraform_docs_generator/openapi_terraform_docs_generator/templates/zendesk"
 	"io"
 )
