+++ utils/terraform_docs_generator/openapi_terraform_docs_generator/openapi_terraform_provider_documentation_data_sources.go
@@ -1,4 +1,4 @@
-package openapi
+package openapi_terraform_docs_generator
 
 import (
 	"io"
