+++ utils/terraform_docs_generator/openapi_terraform_docs_generator/openapi_terraform_provider_documentation_properties.go
@@ -1,4 +1,4 @@
-package openapi
+package openapi_terraform_docs_generator
 
 // Property defines the attributes for describing a given property for a resource
 type Property struct {
