+++ utils/terraform_docs_generator/openapi_terraform_docs_generator/templates/zendesk/zendesk_special_terms_tmpl.go
