+++ utils/terraform_docs_generator/openapi_terraform_docs_generator/utils.go
@@ -1,4 +1,4 @@
-package openapi
+package openapi_terraform_docs_generator
 
 import (
 	"html/template"
